Verion 2.2

+++ app/src/main/java/com/example/olgac/tutor_temp/model/db/TutorDAO.java
@@ -52,7 +52,7 @@ public interface TutorDAO {
 
 
     @Query("DELETE FROM Tutor WHERE tutorId = :tutorID")
-    void deleteTutor(int tutorID);
+    void deleteTutor(long tutorID);
 
     @Query("DELETE FROM Tutor")
     void deleteAll();
